pyproject.toml
[project]
name = "hat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "google-genai>=1.20.0",
]
